♻️ refactor: getAgeIconSrc 분리
